+++ file_reader/readers/__init__.py
@@ -1,3 +1,2 @@
-from .dir_to_df import *
 from .gen_readers import *